% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hearhead.R
\name{hearhead}
\alias{hearhead}
\title{Gives an html table of the first few lines of the data frame along with a description of the variables
and dimension of the data frame.}
\usage{
hearhead(data.frame, nlines = 6, dec.places = 3)
}
\arguments{
\item{data.frame}{a data frame}

\item{nlines}{How many rows of the data frame will be shown in the table (default 6)}

\item{dec.places}{How many decimal places should be shown if the variable is numeric (default 3)}
}
\value{
an html document with a table that has variable names, descriptions, and the first few lines of the data frame
}
\description{
Gives an html table of the first few lines of the data frame along with a description of the variables
and dimension of the data frame.
}
\examples{
data(iris)
hearhead(iris)

}
